vBeta 0.10
New Navbar, yes, I will be fixing the shadows of the buttons in the next update.

About page, basically, a page about me. Theres nothing in it yet. Once there is and I have an intro above the carousel, we will be out of Beta!

Fixed the license on the info page, there was an issue with CSS where it didnt have its own section on the monospace font.